Memorize full metric name to avoid re-construct full metric name
